Moved event loops to dedicated file, fixed selection histogram reference bug

+++ include/Structures.hpp
@@ -41,7 +41,7 @@ struct TruePhotonHit {
 struct TruePhoton {
     TVector3 dir;
     TLorentzVector p4;
-    int parentID;
+    double parentID;
 };
 
 struct TruePi0 {
